+++ spec/interactors/setup_game_spec.rb
@@ -25,15 +25,15 @@ describe SetupGame do
       call
       expect(game.players).to eq(
         [
-          Player.new(name: 'Bartek', color: :orange),
-          Player.new(name: 'Leo', color: :blue)
+          Player.new(index: 0, name: 'Bartek', color: :orange),
+          Player.new(index: 1, name: 'Leo', color: :blue)
         ]
       )
     end
 
     it 'sets up current player' do
       call
-      expect(game.current_player).to eq(Player.new(name: 'Bartek', color: :orange))
+      expect(game.current_player).to eq(Player.new(index: 0, name: 'Bartek', color: :orange))
     end
   end
 
@@ -62,17 +62,17 @@ describe SetupGame do
       call
       expect(game.players).to eq(
         [
-          Player.new(name: 'Bartek', color: :orange),
-          Player.new(name: 'Leo', color: :blue),
-          Player.new(name: 'Carles', color: :white),
-          Player.new(name: 'Gerard', color: :red)
+          Player.new(index: 0, name: 'Bartek', color: :orange),
+          Player.new(index: 1, name: 'Leo', color: :blue),
+          Player.new(index: 2, name: 'Carles', color: :white),
+          Player.new(index: 3, name: 'Gerard', color: :red)
         ]
       )
     end
 
     it 'sets up current player' do
       call
-      expect(game.current_player).to eq(Player.new(name: 'Bartek', color: :orange))
+      expect(game.current_player).to eq(Player.new(index: 0, name: 'Bartek', color: :orange))
     end
   end