+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Test interactors (used to synthetically manipulate state in tests)
 - CurrentPlayerCalculator helper
 - GameSerializer utility
+- index attribute to Player entity
 
 ### Changed
 - ...